product_search/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
qdrant-client==1.7.0
google-generativeai==0.3.2
python-multipart==0.0.6
pyyaml==6.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0